+++ client/src/components/pages/Groups/GroupsRecent.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { Loader, Dimmer, Grid, Header, Segment, Label } from "semantic-ui-react";
+import { Grid, Header, Segment, Label } from "semantic-ui-react";
 import { Link } from 'react-router-dom';
 
 /**
@@ -7,12 +7,9 @@ import { Link } from 'react-router-dom';
  *  request to join a community with the `Join` link. Once clicked, user sees
  *  `Requested` which remains until request is approved or rejected.
  *
- *  @param {boolean} isAuth Determines if user is authenticated
  *  @param {array} groupsActivity Data for active groups
- *  @param {string} groupRequested The name of the group being requested to join
- *  @param {function} onJoinGroup Handles a join request
  */
-const GroupsRecent = ({ isAuth, groupsActivity, groupRequested, onJoinGroup }) => {
+const GroupsRecent = ({ groupsActivity }) => {
   if (groupsActivity.length) {
     return (
       <React.Fragment>
@@ -26,35 +23,19 @@ const GroupsRecent = ({ isAuth, groupsActivity, groupRequested, onJoinGroup }) =
                     <Label attached='top' className='head'>
                       <div className='left'>
                         <Header as='h3'>
-                          {g.display}
+                          <Link
+                            to={`group/${g.name}`}
+                          >
+                            {g.display}
+                          </Link>
                         </Header>
-                      </div>
-                      <div className='right'>
-                        {
-                          (isAuth)
-                          ? (groupRequested === g.name)
-                            ? <Dimmer inverted active><Loader /></Dimmer>
-                            : (g.access && g.access.access !== 100)
-                              ? ''
-                              : (g.access && g.access.access === 100)
-                                ? 'Requested'
-                                : (
-                                  <a
-                                    href={`/join/${g.name}`}
-                                    onClick={e => onJoinGroup(e, g.name)}
-                                  >
-                                    {'Join'}
-                                  </a>
-                                )
-                          : 'Login to Join'
-                        }
                       </div>
                       <div className='clear' />
                     </Label>
                     <ul className='custom-list'>
                       {
-                        g.posts.length
-                        ? g.posts.map(p => (
+                        g.kposts.length
+                        ? g.kposts.map(p => (
                           <li key={p._id}>
                             <Link
                               to={p.st_category+'/@'+p.st_author+'/'+p.st_permlink}
