+++ client/src/components/pages/Manage/GroupManagePosts.js
@@ -1,9 +1,8 @@
 import React from 'react';
-import { Grid, Segment, Icon, Table, Dimmer, Loader } from "semantic-ui-react";
+import { Grid, Segment } from "semantic-ui-react";
 import PropTypes from 'prop-types';
-import { Link } from 'react-router-dom';
 
-import {roles} from '../../../settings';
+import GroupPosts from '../../Common/GroupPosts'
 
 /**
  *  Table of posts for the selected group.
@@ -17,7 +16,8 @@ import {roles} from '../../../settings';
  *  @param {string} props.deletingPost The post to be deleted
  *  @returns {Component} Table of post data
  */
-const GroupManagePosts = ({posts, showModal, deletingPost, user, access, headers}) => {
+const GroupManagePosts = ({posts, showModal, deletingPost, user, access}) => {
+
   if (!posts.length) {
     return (
       <Grid.Column width={8}>
@@ -29,62 +29,13 @@ const GroupManagePosts = ({posts, showModal, deletingPost, user, access, headers
   }else {
     return (
       <Grid.Column>
-        <Table striped>
-          <Table.Header>
-            <Table.Row>
-              {
-                headers.map((h, i) => {
-                  let align = { 'textAlign': 'center' };
-                  if (i === 0) align = {};
-                  return <Table.HeaderCell {...align} key={h}>{h}</Table.HeaderCell>
-                })
-              }
-            </Table.Row>
-          </Table.Header>
-          <Table.Body>
-            {
-            posts.map((p, i) => (
-              <Table.Row key={p._id}>
-                <Table.Cell>
-                <Link
-                  to={p.st_category+'/@'+p.st_author+'/'+p.st_permlink}
-                >
-                {(p.st_title.length > 70)
-                  ? p.st_title.substr(0,70) + " ..."
-                  : p.st_title}
-                </Link>
-                </Table.Cell>
-                <Table.Cell collapsing textAlign='center'>{p.likes}</Table.Cell>
-                <Table.Cell collapsing textAlign='center'>{p.views}</Table.Cell>
-                <Table.Cell collapsing textAlign='center'>{p.rating}</Table.Cell>
-                <Table.Cell collapsing textAlign='center'>{p.added_by}</Table.Cell>
-                <Table.Cell collapsing textAlign='center'>
-                  {
-                    (deletingPost === p.st_permlink)
-                      ? <Dimmer inverted active><Loader /></Dimmer>
-                      : ''
-                  }
-                  {
-                    //if logged in user added post, or is mod or above
-                    //can delete the post
-                    //(user === p.added_by)
-                    //?
-                    access < roles.kGroupsRolesRev['Member']
-                    &&
-                      (
-                        <a href={'/post/delete/'+p.st_author+'/'+p.st_permlink} onClick={e => showModal(e, {author: p.st_author, post: p.st_permlink})}>
-                          <Icon name='delete' color='blue' />
-                        </a>
-                      )
-
-                    //:''
-                  }
-                </Table.Cell>
-              </Table.Row>
-            ))
-          }
-          </Table.Body>
-        </Table>
+        <GroupPosts
+          posts={posts}
+          showModal={showModal}
+          deletingPost={deletingPost}
+          user={user}
+          access={access}
+        />
       </Grid.Column>
     )
   }
