+++ client/src/settings.js
@@ -2,7 +2,7 @@
  *  Common data settings to use.
  */
 
-export const BASE_STEEM_URL = "https://localhost:3000";
+export const BASE_URL = "https://localhost:3000";
 export const roles = {
   kGroupsRoles: {
     0: 'Owner',
@@ -28,4 +28,4 @@ export const roles = {
   }
 }
 
-export default BASE_STEEM_URL;
+export default BASE_URL;
