+++ client/src/utils/auth/scAPI.js
@@ -6,8 +6,8 @@ import scConfig from './scConfig'
  */
 const api = sc2.Initialize({
   app: scConfig.sc.app,
-  callbackURL: scConfig.sc.callbackURLDev,
-  //callbackURL: process.env.NODE_ENV === 'development' ? 'http://localhost:3000/callback' : 'https://appbaseio-apps.github.io/reactivesearch-auth0-example/callback',
+  //callbackURL: scConfig.sc.callbackURLDev,
+  callbackURL: process.env.NODE_ENV === 'development' ? scConfig.sc.callbackURLDev : scConfig.sc.callbackURLProd,
   accessToken: scConfig.sc.accessToken,
   scope: scConfig.sc.scope,
 });
