+++ pm2.config.js
@@ -1,0 +1,6 @@
+module.exports = {
+  apps: [{
+    name: 'kure',
+    script: './start-server.js',
+  }]
+}
