+++ server/routes/api/api.js
@@ -1,6 +1,8 @@
 import { Router } from 'express';
 import groups from './groups';
 import recentActivity from './recentActivity';
+import auth from './auth/auth';
+import manage from './manage/manage';
 //const initialState = serialize(response);
 //var html = xss('<script>alert("xss");</script>');
 
@@ -8,5 +10,7 @@ import recentActivity from './recentActivity';
 const router = new Router();
 router.use('/groups', groups);
 router.use('/recentactivity', recentActivity);
+router.use('/auth', auth);
+router.use('/manage', manage);
 
 export default router;
